+++ lf/system/admin/controller/navigation.php
@@ -255,157 +255,4 @@ class navigation
 		
 		return $select;
 	}
-
-	public function preview()
-	{
-		$param = \lf\requestGet('param');
-		// generate normal frontend
-		
-		(new \lf\cms)->loadLfCss();
-		
-		// load in nav cache
-		$previewNav = (new \lf\cms)->getNavCache();
-		//$replace = [
-			// '%baseurl%' => \lf\requestGet('AdminUrl').'navigation/',
-			// '<a ' => '<a target="_parent"'
-		// ];
-		//$previewNav = str_replace(array_keys($replace), array_values($replace), $navCache);
-		
-		
-		
-		$replace = [
-			'%baseurl%' => \lf\requestGet('AdminUrl').'navigation/',
-			'<a ' => '<a target="_parent"'
-		];
-		
-		$previewNav = str_replace(array_keys($replace), array_values($replace), $previewNav);
-		
-		
-		
-		// new request
-		$request = (new \lf\request)->load()
-			// Drop navigation into Cwd (doesnt affect anything really...)
-			->actionDrop()
-			// set param as what $this controller method received
-			->paramShift(2)
-			//->setAction([])
-			->save();
-		
-		(new \lf\cms)->getContent($param[1]);
-		
-		$rendered = (new \lf\template)
-			->addContent($previewNav, 'nav')
-			->setAdmin(false)
-			->render();
-			
-		
-		//$rendered = str_replace(array_keys($replace), array_values($replace), $rendered);
-		echo $rendered;
-			
-		
-	//	pre(new \lf\template, 'var_dump');
-		
-			// replace /admin/navigation/preview/ into template nav output
-		// set request to frontend mode and execute apps (getcontent?)
-		
-		
-		// print template
-		
-		
-		
-		exit;
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		// never gets here
-		
-		
-		return;
-		
-		$vars =  \lf\requestGet('Param');
-		
-		$action = (new \LfActions)->findById($vars[1]);
-		$links = (new \LfLinks)->findByInclude($vars[1]);
-		
-		
-		$skin = $action->template;
-		if($skin == 'default')
-			$skin = \lf\getSetting('default_skin');
-		
-		
-		//(new \lf\cms)->getContent();
-		
-		ob_start();
-		include LF.'cache/nav.cache.html';
-		$nav = ob_get_clean();
-		
-		(new \lf\template)->addContent('nav', $nav);
-		
-		ob_start();
-		include LF.'skins/'.$skin.'/index.php';
-		$template = ob_get_clean();
-		
-		$content = '<h2>%content%</h2>';
-		
-		
-		//pre($links->result);
-		
-		$content .= implode(', ', $action->get()).'<br />';
-		
-		foreach($links->result as $row)
-		{
-			$content .= implode(', ', $row).'<br />';
-		}
-		
-		
-		
-		$content .= '
-			<div class="row">
-				<div class="col-4">Add new app:</div>
-				<div class="col-4">
-					<select name="" id="">
-						<option value="">App1</option>
-					</select>
-				</div>
-				<div class="col-4">
-					<input type="submit" />
-				</div>
-			</div>
-		 
-		
-		';
-		
-		
-		
-		$template = str_replace(
-			array(
-				'%content%',
-				'%skinbase%',
-				'%nav%',
-				'%baseurl%',
-				'</head>'
-			),
-			array(
-				$content,
-				\lf\requestGet('LfUrl').'skins/'.$skin,
-				$nav,
-				\lf\requestGet('AdminUrl').'dashboard/preview/'.$vars[1].'/',
-				'<link rel="stylesheet" href="'.\lf\requestGet('LfUrl').'system/lib/lf.css" /><link rel="stylesheet" href="'.\lf\requestGet('LfUrl').'system/lib/3rdparty/icons.css" /></head>'
-			),
-			$template
-		);
-		
-		echo $template;
-		
-		exit();
-	}
 }
